fix includes for cuda library headers
The problem with CUDA compilation was a number of transitively recursive includes made to cuda-all.h
inside the library headers.

This caused the preprocessor to stop #include expansion early, leaving out many key definitions in
the different CUDA headers.

All references to cuda-all.h in headers was removed for this reason. Compilation is now successful