remove perform_clipping from sum_prod_alg_inf
Clipping is only performed internally by the algorithm, and hence it does not make sense for this method to be a part of the interface, especially when it gets in the way of a GPU implementation of the SPA, for example.

Note that the impl. in sum_prod_alg_abstract was also declared virtual to maintain some of the flexibility of the previous implementation.